feat: change Training Points from int to long to prevent overflow issues
- Updated the TP storage type across the system from int to long to accommodate larger values and prevent overflow.
- Implemented a new TPNumberFormatter utility for displaying TP values with appropriate suffixes (B, T, Q).
- Ensured backward compatibility for existing saves by checking the NBT type during loading.
- Adjusted relevant commands and events to handle the new long type for TP calculations and manipulations.

+++ src/main/java/com/dragonminez/client/gui/MastersSkillsScreen.java
@@ -210,7 +210,7 @@ public class MastersSkillsScreen extends BaseMenuScreen {
 
 		if (!statsData.getSkills().hasSkill(selectedSkill) || skill.getLevel() == 0) {
 			int cost = getUpgradeCost(selectedSkill, 0);
-			int currentTPS = statsData.getResources().getTrainingPoints();
+			long currentTPS = statsData.getResources().getTrainingPoints();
 			boolean canAfford = currentTPS >= cost;
 			if (cost == -1 || cost == Integer.MAX_VALUE) return;
